added RULES and NOTINLINE and INLINE pragmas
Hopefully this will keep code size relatively small while still allowing partial
evaluation via RULES.

For functions that don't depend on the type of the Set, I've used NOINLINE in
Internal and INLINE in Word64 and Int64. For functions that do depend on the type, I've used INLINE in Internal
and NOTINLINE in Word64 and Int64.